feat: add view bids to modal provider and refactor existing code
