+++ src/containers/other/NFTView/elements/Actions/Actions.tsx
@@ -11,9 +11,6 @@ import {
 // Library
 import { formatEther } from '@ethersproject/units';
 import { Bid } from '@zero-tech/zns-sdk/lib/zAuction';
-import { DisplayParentDomain, Maybe, Metadata } from 'lib/types';
-import { Domain } from '@zero-tech/zns-sdk/lib/types';
-import { ethers } from 'ethers';
 
 //- Types Imports
 import { ActionBlock, ACTION_TYPES } from './Actions.types';
@@ -23,12 +20,6 @@ import { LABELS } from 'constants/labels';
 import { CURRENCY } from 'constants/currency';
 import { TEST_ID, wrapFiat } from './Actions.constants';
 
-//- NFTView Hooks Imports
-import { useNFTViewModal } from '../../hooks';
-
-//- Modal Provider Imports
-import { NFTViewModalType } from '../../providers/NFTViewModalProvider/NFTViewModalProvider.types';
-
 // Styles
 import styles from './Actions.module.scss';
 import classNames from 'classnames/bind';
@@ -37,48 +28,31 @@ const cx = classNames.bind(styles);
 
 type ActionsProps = {
 	domainId?: string;
-	makeABidDomainData?: Maybe<DisplayParentDomain>;
-	viewBidsDomainData?: Domain;
-	domainMetadata?: Maybe<Metadata>;
 	buyNowPrice?: number;
 	highestBid?: number;
 	onMakeBid: () => void;
+	onViewBids: () => void;
 	yourBid?: Bid;
 	isBiddable?: boolean;
 	isOwnedByUser?: boolean;
 	wildPriceUsd?: number;
 	refetch: () => void;
 	bidData?: Bid[];
-	isLoading?: boolean;
-	setIsViewBidsOpen?: (state: boolean) => void;
-	isViewBidsOpen?: boolean;
-	setIsSetBuyNowOpen?: (state: boolean) => void;
-	isSetBuyNowOpen?: boolean;
 };
 
 const Actions = ({
 	domainId,
-	domainMetadata,
-	makeABidDomainData,
-	viewBidsDomainData,
 	buyNowPrice,
 	highestBid,
 	onMakeBid,
+	onViewBids,
 	yourBid,
 	isBiddable,
 	isOwnedByUser,
 	wildPriceUsd,
 	refetch,
 	bidData,
-	isLoading,
-	setIsViewBidsOpen,
-	isViewBidsOpen,
-	setIsSetBuyNowOpen,
-	isSetBuyNowOpen,
 }: ActionsProps) => {
-	//- Modal Provider
-	const { openModal, closeModal } = useNFTViewModal();
-
 	//- Condition helpers
 	const isBidData = bidData && bidData?.length > 0;
 	const isYourBids = !isOwnedByUser && Boolean(yourBid);
@@ -87,10 +61,6 @@ const Actions = ({
 	const isBuyNow = Boolean(buyNowPrice) && !isOwnedByUser && Boolean(domainId);
 	const isViewBids =
 		isOwnedByUser !== undefined && isBiddable === true && Boolean(isBidData);
-
-	// Convert highest bid as wei
-	const highestBidAsWei =
-		highestBid && ethers.utils.parseEther(highestBid.toString()).toString();
 
 	const highestBidTextValue =
 		Boolean(highestBid) && Boolean(wildPriceUsd)
@@ -112,30 +82,6 @@ const Actions = ({
 			? wrapFiat(buyNowPrice * wildPriceUsd)
 			: LABELS.NO_BUY_NOW;
 
-	///////////////
-	// Functions //
-	///////////////
-
-	const handleOnBid = () => {
-		onMakeBid();
-		closeModal();
-	};
-
-	// Open Make A Bid Modal
-	const openMakeABid = () => {
-		openModal({
-			modalType: NFTViewModalType.MAKE_A_BID,
-			contentProps: {
-				domain: makeABidDomainData!,
-				onBid: handleOnBid,
-				onClose: closeModal,
-			},
-		});
-	};
-
-	/**
-	 * This is a pretty messy structure - could be refactored moving forward
-	 */
 	const actions: { [action in ACTION_TYPES]: ActionBlock } = {
 		[ACTION_TYPES.BuyNow]: {
 			amount: buyNowPrice,
@@ -164,8 +110,6 @@ const Actions = ({
 					domainId={domainId ?? ''}
 					isTextButton={isTextButton}
 					className={cx({ TextButton: isTextButton })}
-					setIsSetBuyNowOpen={setIsSetBuyNowOpen}
-					isSetBuyNowOpen={isSetBuyNowOpen}
 				/>
 			),
 			isVisible: isSetBuyNow,
@@ -181,22 +125,15 @@ const Actions = ({
 						className={cx({ TextButton: isTextButton })}
 						isTextButton={isTextButton}
 						glow
-						onClick={openMakeABid}
+						onClick={onMakeBid}
 					>
 						{placeBidButtonTextValue}
 					</BidButton>
 				) : (
 					<ViewBidsButton
-						bids={bidData}
-						domain={viewBidsDomainData}
 						isTextButton={isTextButton}
-						highestBidAsWei={String(highestBidAsWei)}
-						refetch={refetch}
-						isLoading={isLoading}
+						onClick={onViewBids}
 						className={cx({ TextButton: isTextButton })}
-						domainMetadata={domainMetadata}
-						setIsViewBidsOpen={setIsViewBidsOpen}
-						isViewBidsOpen={isViewBidsOpen}
 					/>
 				);
 			},
