+++ src/containers/other/NFTView/elements/NFT/NFT.tsx
@@ -11,19 +11,8 @@ import styles from './NFT.module.scss';
 
 //- Library Imports
 import { truncateWalletAddress } from 'lib/utils';
-import { Domain } from '@zero-tech/zns-sdk/lib/types';
-import { Bid } from '@zero-tech/zauction-sdk';
-import { Maybe, Metadata } from 'lib/types';
-
-//- Constants Imports
-import { NFT_MORE_ACTIONS_TITLE } from '../../NFTView.constants';
-
 //- Type Imports
 import { Option } from 'components/Dropdowns/OptionDropdown/OptionDropdown';
-
-//- NFTView Modal Provider Imports
-import { useNFTViewModal } from '../../hooks';
-import { NFTViewModalType } from '../../providers/NFTViewModalProvider/NFTViewModalProvider.types';
 
 export const Amount = (amount: string) => (
 	<span className={styles.Amount}>{amount}</span>
@@ -35,108 +24,28 @@ type OptionType = {
 }[];
 
 type NFTProps = {
-	bids: Bid[];
-	domainMetadata?: Maybe<Metadata>;
 	owner: string;
 	title: string;
-	domainId: string;
 	assetUrl: string;
 	creator: string;
 	options: OptionType;
-	wildPriceUsd: number;
 	description?: string;
-	isLoading?: boolean;
-	highestBid?: string;
-	viewBidsDomainData?: Domain;
-	refetch: () => void;
-	onTransfer: () => void;
+	onSelectOption: (option: Option) => void;
 	onDownload?: () => void;
 	onShare?: () => void;
 };
 
 const NFT = ({
-	bids,
-	domainMetadata,
 	owner,
 	title,
-	domainId,
 	assetUrl,
 	creator,
 	options,
-	wildPriceUsd,
 	description,
-	isLoading,
-	highestBid,
-	viewBidsDomainData,
-	refetch,
-	onTransfer,
+	onSelectOption,
 	onDownload,
 	onShare,
 }: NFTProps) => {
-	//- Modal Provider Hook
-	const { openModal, closeModal } = useNFTViewModal();
-
-	///////////////
-	// Functions //
-	///////////////
-
-	// Open Domain Settings Modal
-	const openDomainSettings = () => {
-		openModal({
-			modalType: NFTViewModalType.DOMAIN_SETTINGS,
-			contentProps: {
-				domainId: domainId,
-				onClose: closeModal,
-			},
-		});
-	};
-
-	// Open Domain Settings Modal
-	const openSetBuyNow = () => {
-		openModal({
-			modalType: NFTViewModalType.SET_BUY_NOW,
-			contentProps: {
-				domainId: domainId,
-				onCancel: closeModal,
-				onSuccess: refetch,
-			},
-		});
-	};
-
-	const handleOnAccept = () => {
-		refetch();
-		closeModal();
-	};
-
-	// Open Bid List Modal
-	const openBidList = () => {
-		openModal({
-			modalType: NFTViewModalType.BID_LIST,
-			contentProps: {
-				bids: bids,
-				domain: viewBidsDomainData,
-				domainMetadata: domainMetadata ?? undefined,
-				onAccept: handleOnAccept,
-				wildPriceUsd,
-				isLoading,
-				highestBid,
-			},
-		});
-	};
-
-	// Dropdown Option Select
-	const onSelectOption = (option: Option) => {
-		if (option.title === NFT_MORE_ACTIONS_TITLE.MY_DOMAIN_SETTINGS) {
-			openDomainSettings();
-		} else if (option.title === NFT_MORE_ACTIONS_TITLE.TRANSFER_OWNERSHIP) {
-			onTransfer();
-		} else if (option.title === NFT_MORE_ACTIONS_TITLE.SET_BUY_NOW) {
-			openSetBuyNow();
-		} else {
-			openBidList();
-		}
-	};
-
 	////////////
 	// Render //
 	////////////
